requirements.txt
scikit-learn
numpy
pandas
flask
python-dotenv
matplotlib
pymongo[srv]
pymongo
certifi
pyaml
dill
mlflow
dagshub
fastapi
uvicorn
#-e .